removed signed/unsigned comparison warning
git-svn-id: file:///usr/groups/dtg/svn/Cantag/trunk@1746 02f9cdae-41fc-0310-9451-d5e721ae9d13

+++ src/Camera.cc
@@ -363,8 +363,8 @@ namespace Cantag {
 
   void Camera::UnDistortImage(Image<Pix::Sze::Byte1,Pix::Fmt::Grey8>& image) const {
     const Image<Pix::Sze::Byte1,Pix::Fmt::Grey8> source(image);
-    for(short i=0;i<image.GetHeight();i++) {
-      for(short j=0;j<image.GetWidth();j++) {
+    for(unsigned int i=0;i<image.GetHeight();i++) {
+      for(unsigned int j=0;j<image.GetWidth();j++) {
 	float points[] = {i,j};
 	points[0] -= m_intrinsic[2];
 	points[1] -= m_intrinsic[5];
